+++ src/doctor/tests.rs
@@ -15,6 +15,7 @@ use super::schema::{check_key_format, check_schema_migration};
 mod health_action_tests;
 mod log_health;
 mod memory_usage_feedback;
+mod pack_imports;
 mod promotion_funnel;
 
 struct ScopedCipherKeyEnv {